Allow orgId is empty

+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/IntervenerService.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.reform.finrem.caseorchestration.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.ObjectUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.finrem.caseorchestration.handler.FinremCallbackRequest;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.FinremCaseData;
@@ -115,7 +116,7 @@ public class IntervenerService {
             && beforeIntv.getIntervenerRepresented() != null
             && beforeIntv.getIntervenerRepresented().equals(YesOrNo.YES)) {
             String beforeOrgId = beforeIntv.getIntervenerOrganisation().getOrganisation().getOrganisationID();
-            if (!beforeOrgId.equals(orgId) || !beforeIntv.getIntervenerSolEmail().equals(email)) {
+            if (ObjectUtils.notEqual(beforeData, orgId) || !beforeIntv.getIntervenerSolEmail().equals(email)) {
                 revokeIntervenerRole(caseDetailsBefore.getId(), beforeIntv.getIntervenerSolEmail(),
                     beforeOrgId,
                     intervenerWrapper.getIntervenerSolicitorCaseRole().getCcdCode(), errors);
@@ -159,7 +160,6 @@ public class IntervenerService {
         intervenerWrapper.getIntervenerOrganisation().setOrgPolicyReference(null);
     }
 
-
     public IntervenerChangeDetails setIntervenerAddedChangeDetails(IntervenerWrapper intervenerWrapper) {
         IntervenerChangeDetails intervenerChangeDetails = new IntervenerChangeDetails();
         intervenerChangeDetails.setIntervenerAction(IntervenerAction.ADDED);
@@ -180,7 +180,12 @@ public class IntervenerService {
     }
 
     private void addIntervenerRole(Long caseId, String email, String orgId, String caseRole, List<String> errors) {
+        if (StringUtils.isBlank(orgId)) {
+            // skip this logic if orgId is not set
+            return;
+        }
         Optional<String> userId = organisationService.findUserByEmail(email, systemUserService.getSysUserToken());
+
         if (userId.isPresent()) {
             assignCaseAccessService.grantCaseRoleToUser(caseId, userId.get(), caseRole, orgId);
         } else {
@@ -189,6 +194,10 @@ public class IntervenerService {
     }
 
     private void revokeIntervenerRole(Long caseId, String email, String orgId, String caseRole, List<String> errors) {
+        if (StringUtils.isBlank(orgId)) {
+            // skip this logic if orgId is not set
+            return;
+        }
         Optional<String> userId = organisationService.findUserByEmail(email, systemUserService.getSysUserToken());
         if (userId.isPresent()) {
             assignCaseAccessService.removeCaseRoleToUser(caseId, userId.get(), caseRole, orgId);
